refactor: new folder structure to scale sdk
