+++ src/constants/internal.ts
@@ -1,0 +1,5 @@
+import { BigNumber as BN } from "ethers";
+
+export const ZERO = BN.from("0")
+export const ONE = BN.from("1")
+export const ZERO_ADDRESS = "0x0000000000000000000000000000000000000000";
