+++ src/constants/stake.ts
@@ -1,0 +1,13 @@
+export const XSLOCKER_ADDRESS: { [chainID: number]: string } = {
+    [1]: "0x501Ace47c5b0C2099C4464f681c3fa2ECD3146C1",
+    [4]: "0x501Ace47c5b0C2099C4464f681c3fa2ECD3146C1",
+    [137]: "0x501Ace47c5b0C2099C4464f681c3fa2ECD3146C1",
+    [1313161554]: "0x501Ace47c5b0C2099C4464f681c3fa2ECD3146C1"
+}
+
+export const STAKING_REWARDS_ADDRESS: { [chainID: number]: string } = {
+    [1]: "0x501ace3D42f9c8723B108D4fBE29989060a91411",
+    [4]: "0x501ace3D42f9c8723B108D4fBE29989060a91411",
+    [137]: "0x501ace3D42f9c8723B108D4fBE29989060a91411",
+    [1313161554]: "0x501ace3D42f9c8723B108D4fBE29989060a91411"
+}
