+++ src/constants/token.ts
@@ -1,0 +1,13 @@
+export const SOLACE_ADDRESS: { [chainID : number]: string } = {
+    [1] : "0x501acE9c35E60f03A2af4d484f49F9B1EFde9f40", // Eth Mainnet
+    [4] : "0x501acE9c35E60f03A2af4d484f49F9B1EFde9f40", // Rinkeby
+    [137] : "0x501acE9c35E60f03A2af4d484f49F9B1EFde9f40", // Polygon
+    [1313161554] : "0x501acE9c35E60f03A2af4d484f49F9B1EFde9f40", // Aurora
+}
+
+export const DAI_ADDRESS: { [chainID : number]: string } = {
+    [1] : "0x6B175474E89094C44Da98b954EedeAC495271d0F",
+    [4] : "0x8ad3aA5d5ff084307d28C8f514D7a193B2Bfe725",
+    [137] : "0x8f3Cf7ad23Cd3CaDbD9735AFf958023239c6A063",
+    [1313161554] : "0xe3520349F477A5F6EB06107066048508498A291b",
+}
