+++ src/index.ts
@@ -1,7 +1,6 @@
 export * from "./constants"
-export * from "./fetcher"
-export * from "./policyholder"
-export * from "./staker"
-export * from "./bonder"
+export * from "./contracts"
+export * from "./apis"
+export * from "./types"
 export * from "./utils"
 export * from 'ethers'
